waf: Add option to align ZMQ quirk with CSPv1
Introduce a new Waf option `--enable-fixup-v1-zmq-little-endian`. When
enabled, this option applies a fixup to the ZMQ interface to match the
libcsp v1 behavior, which incorrectly used host endianness instead of
the expected big-endian format.

This fixup ensures compatibility with systems using libcsp v1 while
aligning with the behavior of libcsp v2.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>