csp_types: Ensure csp_id_t structure is properly packed
The `csp_id_t` structure was not actually packed since the `__packed`
macro was not defined in `csp_types.h`.

`__packed` is defined in `csp_macro.h`, which is located in the `src/`
directory and is not included (nor accessible) from
`include/csp/csp_types.h`.

As a result, the compiler ignored the packing directive, potentially
leading to unexpected padding.

Replaced `__packed` with `__attribute__((__packed__))` to explicitly
enforce packing.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>