Handle packets to be routed forward and those for the itf owner differently
In the current csp_can_pbuf_new() function implementation, CSP buffer exhaustion will cause the calling thread to crash (csp_buffer_get_always() will assert).
In the context of a CSH instance acting as a router, this means that as soon as this situation occurs, the router will stop functioning *silently*,
requiring a manual restart of CSH (the router thread will enter an infinite loop due to csp_panic() not being over-written in CSH (separate issue to be addressed)).
We have observed this behaviour where bursts of packets to be routed will cause the router to stop working due to buffer exhaustion.
To mitigate this issue, csp_can_pbuf_new() was changed to handle buffer exhaustion more gracefully, dropping to be routed packets, while still aborting in case
where the packet was for the interface owner, but not buffers where available.
Dropping packets will allow code to free up CSP buffers once the burst is over, making the infrastructure much more resilient.

The same logic is applied in ZMQ's csp_zmqhub_task() function